Fixed errors from moving files
